Move `startstop` package up to root of `./internal/` (#419)
A very small change that's been on my list for a while, but I didn't do
because it'd be conflicting with some other branches I had: move the
`startstop` package up to root of `./internal/`. Until now it's been in
`./internal/maintenance` because it was originally only used for
maintenance services, but since it's now used much more widely, it makes
more sense for it to live in `./internal`.