Prepare release 0.9.0 + add test for periodic job bundle
We haven't cut a new release in a while, and the bug in #420 is kind of
bad, so now that there's a fix it's not a bad time to cut a new release.

I'm also including a test case that catches the bug from #420.
Apparently none of our other test cases can catch the problem, so it's
good to have a regression guard.