Use longer `WaitTimeout` for test manager pool acquisition context timeout
I'm not sure what changed exactly, but about 25% of my builds have been
failing for the last couple days on the context deadline set when
acquiring a test database.

Here, attempt to resolve the problem by changing the hard-coded 5 second
wait timeout to acquire a test database manager to the timeout specified
by `WaitTimeout`, which is a tight 3 seconds locally, but which allows
more time for GitHub Actions, which seems to regularly be quite slow.

    func WaitTimeout() time.Duration {
        if os.Getenv("GITHUB_ACTIONS") == "true" {
                return 10 * time.Second
        }

        return 3 * time.Second
    }