Remove component status infrastructure in favor of reusing start/stop (#436)
We've made quite a few changes in recent months around winding
stop/start into everything to the point where all services are now
compliant with this concept, including the main client.

Quite a long time ago we put in a "component monitor" framework to help
tests wait for clients to successfully start up. It worked reasonably
well, but is a large amount of code, not very generic, and involves
winding callback functions into all components, which is a little
unsightly.

Here, I'm proposing that we replace the job component monitor was doing
with a more simple wait on the `Started` channel of the client:

    startClient(ctx, t, client)
    riversharedtest.WaitOrTimeout(t, client.baseStartStop.Started())

This is made possible because similar to how the component monitor
framework worked, the client doesn't signal that it's fully started
until it's successfully waited for all its subservices to start up:

    func (c *Client[TTx]) Start(ctx context.Context) error {
        ...

        go func() {
            // Wait for all subservices to start up before signaling our own start.
            // This isn't strictly needed, but gives tests a way to fully confirm
            // that all goroutines for subservices are spun up before continuing.
            //
            // Stop also cancels the "started" channel, so in case of a context
            // cancellation, this statement will fall through. The client will
            // briefly start, but then immediately stop again.
            startstop.WaitAllStarted(append(
                c.services,
                producerServices..., // see comment on this variable
            )...)

            started()
            defer stopped()

So all in all, we should be able to get similar guarantees for test
purposes while needing less code and getting more reusability (in that
non-client services can also use exactly the same code to wait for start
up).