Have client's `TestOnly` activate `AsyncCompleter` instead of `BatchCompleter`
I was looking into why River's example tests were so incredibly slow. By
far the biggest reason was that maintenance services were jittering on
start up, but by far the _second_ biggest reason is that the batch
completer pauses before carrying out any completion (waiting to see if
anymore jobs will come in to add to the completion batch), which delays
subscribers from receiving a confirmation that a job's been finished.

Here, I propose that we have the `TestOnly` flag activate the async
completer instead of the normal batch completer. This removes the pause
mentioned above, and especially in test cases where job volume is
expected to be low, will have no other performance disadvantage. The
batch completer is generally more complicated, so this will probably
also lower memory overhead.

There is some risk in removing the batch completer from standard test
usage in that it means less load on the batch completer, but I think is
probably okay because the batch completer's test suite in isolation is
very comprehensive.

An alternative to this is might be to make the client's completer
configurable through `Config`. I thought about that approach, but it has
the downside of exposing more API, and the async completer will be more
appropriate for use in 99.9% of test cases anyway, so it's an extra
complication with no real advantage.