Demote many log lines from info to debug
We've had reports that a running River client can be quite a noisy
affair, producing too much logging in general.

Here, we move many log lines from info to debug. I suspect the most
impactful is that we had a bunch in the elector and notifier that would
fire quite often during run of the mill operation.

I left a couple in place, although I'm not 100% sure it's the right
thing:

* The 5 second heartbeat loop in the client that prints job statistics.
  I figure that if this is the _main_ log line being emitted, with
  others being quite rare, it might be okay. We could also consider
  moving it to be more infrequent like every 10 seconds.

* I left the maintenance service "done run" log lines as info for the
  time being, but I changed them so that they only print in case there
  were >0 results, so I think the result is that most of them are silent
  most of the time, except perhaps in an actual production situation.

I figure we can try this as a pass, see how it is, and then possibly
demote more if it's still annoying.