Upgrade golangci-lint from 1.56 to 1.59
We're now a few versions behind the latest golangci-lint, and given that
it moves pretty fast and upgrading too many versions at once can be a
bit of a job, here, bump us up to version 1.59 to get us back latest.

Fix a couple breakages, but relatively few of them. A couple linters
that we had disabled were renamed, so fix those in configuration to
continue to disable them because they continue to be awful (`err113` and
`mnd`).

There are a couple useful linters brought in for Go 1.22:

* `copyloopvar`: Encourages code to _not_ capture variables in loops
  (e.g. `for ...; i := ; ...`) because as of 1.22, it's no longer
  necessary.

* `intrange`: Encourages looping by integer to be written with `range`
  like `for i := range(5)` instead of the C-style `for i := 0; i < 5; i++`.

I like both of these, but we're still technically trying to support Go
1.21, so I've disabled them for the time being. Once we drop 1.21 (which
could plausibly be as close as a few weeks away as 1.23 is expected
soon), we can turn them back on.