Remove use of `filepath` for interacting with embedded migration files
This one is in pursuit of resolving #482 in which migrations can't
currently run on Windows.

As @bgentry points out regarding the `embed` package [1], where forward
slashes are used for all files regardless of underlying operation system:

> The patterns are interpreted relative to the package directory
> containing the source file. The path separator is a forward slash, even
> on Windows systems.

But `filepath` uses OS-specific separators, making commingling use of
these two packages not advisable:

> The filepath package uses either forward slashes or backslashes,
> depending on the operating system. To process paths such as URLs that
> always use forward slashes regardless of the operating system, see the
> path package.

Here, remove use of `filepath` for migrations by using the `IsDir`
function of a file entry instead, which also turns out to be much
cleaner anyway.

I don't have a Windows system to test on, but this seems like a pretty
likely fix. A Windows-specific job to test the CLI is in progress as
part of #483, which would help prevent future regressions.

Fixes #482.

[1] https://pkg.go.dev/embed#hdr-Directives