Account for schema when checking whether a column exists
This one's aimed at addressing #503, in which migrations won't run when
using an alternate schema with `search_path` because the `ColumnExists`
query isn't accounting for schema in any way.

Here, modify `ColumnExists` so that it looks up within `CURRENT_SCHEMA`,
which will respect search path. Add test assertions to verify that this
works as expected, and similar ones in `TableExists` to avoid any
regressions there (it already does respect search path as written, but a
future change could potentially break that).

Fixes #503.