database/sql: default to current time if scheduled_at unspecified (#504)
* test to demonstrate InsertManyTx ScheduledAt default value

* database/sql: fix InsertManyTx default scheduled_at

The database/sql driver had a slightly different behavior than the pgx
driver in that it was setting the `scheduled_at` time to `null`, rather
than setting the current timestamp if one was not specified.

Additionally, the driver test suite did not cover this case at all.

Fix both of these issues. Fixes #502.