Move `TestSignal` into `rivershared` for use outside of main project
The `TestSignal` paradigm might not be perfect, but as I was writing a
test yesterday that had to involve a bespoke test channel involving many
lines of unsightly and somewhat brittle test-only code, I was reminded
that it's better than the alternative.

Here, move `TestSignal` into `rivershared` so that it can be used
outside of the main River project.