Add Go 1.23 to build matrix + update toolchain (#533)
Add Go 1.23 to the build matrix and update toolchain in `go.work` and
all `go.mod` files. Add development instructions for how to use the new
`make update-mod-go` to do so in an automated way.

I didn't drop Go 1.21 support yet because I figure there isn't a hurry
in jettisoning it immediately, although, we'll have to see how the build
stability is with the expanded matrix.