Update indirect dependencies from `update-mod-version` script
This is driven by [1] and the same bug we'd noticed in another project
previously too. Until now, `update-mod-version` hadn't been updating
versions on indirect dependencies, and `rivershared` in the CLI had been
indirect, thereby causing it to try and resolve v0.11.2 instead of
v0.11.3.

Here, update the script so that indirect dependencies are updated as
well.

[1] https://github.com/riverqueue/river/issues/538