Bump `rivershared` to v0.11.3 in the CLI (#541)
Related to #538, bump `rivershared` to v0.11.3 in the CLI. Instead of
cutting a brand new release here, I think what we can probably do is
swap the CLI's release timestamp with a new one with the `rivershared`
dependency fixed.

There is a chance that that doesn't work because the Go sum database
will prevent a change in checksum. If that's the case, we'll have to
roll forward to v0.11.4 instead.