Prepare release v0.11.4
Prepare release v0.11.4 to fix the bad release in v0.11.3 that was
causing the CLI to become uninstalled as described in #538. The release
script has now been updated in #534 so that the `rivershared` dependency
also gets properly bumped.

I'm also going to try to the `[skip ci]` tag during this release to
hopefully prevent bad caching in the Go proxy and make the release
available sooner.

Fixes #538.

[skip ci]