Upgrade to golangci-lint 1.60.1 (#548)
Here, take the project to golangci-lint 1.60.1, which is now widely
available on Homebrew and elsewhere. It comes with a number of breaking
changes, so I figure it's not the worst idea to keep us on latest so
that future upgrades are less painful.

All breakages are related to overflow checking in `gosec` getting more
strict. The strictness is kind of useful in that it reminds us to check
bounds with something like `min(limit, math.MaxInt32)`, but it can't
detect that a bounds check is done with `min(...)`, which is kind of
annoying because it means you need a `nolint` tag on all of these.