also clear unique_key when should_cancel
While doing some work on a related query, I noticed that the
`unique_key` is being cleared here specifically if the new state is
explicitly `cancelled`, but _not_ if the state is being set to cancel
because of a cancellation attempt that didn't get picked up before the
job completed.

I didn't yet add test coverage for this but wanted to at least commit it
before I forgot.