Add middleware system for jobs
Here, implement a middleware system that adds middleware functions to
job lifecycles, which results in them being invoked during specific
phases of a job like as it's being inserted or worked.

The most obvious unlock for this is telemetry (e.g. logging, metrics),
but it also acts as a building block for features like encrypted jobs.

There are two middleware interfaces added: `JobInsertMiddleware` and
`WorkerMiddleware`. A user could implement these on the same struct type
for something like telemetry where you want to inject a trace ID at job
insertion, and then resurrect it into the context at execution time. In
other cases, such as execution-specific logic, it may only make sense to
implement one of these. Each of these interfaces has a corresponding
`*Defaults` struct that can be embedded to provide future-proofing if
additions are made to them for future functionality.

The `Worker[T]` type has been extended with a `Middleware()` method that
allows each individual worker type to define its own added lists of
middleware which will be run _after_ the global `WorkerMiddleware` at
the client config level. There are also global `JobInsertMiddleware` on
the client config.

Co-authored-by: Blake Gentry <blakesgentry@gmail.com>